requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
groq
python-dotenv

# Para tests
pytest
pytest-asyncio
respx # Para mockear httpx